jsonschema2sql: Move package to lib/core/backend
Change-type: patch
